### 2024-03-21 #6512 @editorialbot
[PRE REVIEW]: Neurosift: DANDI exploration and NWB visualization in the browser
**Submitting author:** @magland (<a href="http://orcid.org/0000-0002-5286-4375">Jeremy Magland</a>)
**Repository:** https://github.com/flatironinstitute/neurosift
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.2.5
