### 2024-03-21 #6512 @Kevin-Mattheus-Moerman
@magland thanks for this submission to JOSS. I am the AEiC on this track. From the above software report it looks like the core functionality of this submission is captured by the `TypeScript` code. For some submissions we've seen that a significant portion of such TypeScript or JavaScript code has been automatically generated, e.g. by a tool facilitating GUI/web interface creation. Can you please clarify what portion e.g. of the 19687 lines of code is your original and personal contribution, and what portion is perhaps auto-generated or taken from existing libraries? Thanks.
