### 2024-03-21 #6512 @magland
Hi @Kevin-Mattheus-Moerman! The vast majority of the code was written by myself or my co-developer Jeff Soules without the assistance of AI. In the past 4 months or so I have been using GitHub Copilot, but this is only for code completion (not writing large sections). In terms of percentages, I would say 90% was written by me and Jeff before copilot assistance and around 10%, more recently, with copilot as an assistant. Essentially no code has been taken from other libraries except for some that was ported over from a related project of mine called [sortingview](https://github.com/magland/sortingview).
