### 2024-03-21 #6512 @magland
> magland if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

After a searching the reviewers list I came upon the following names as potential reviewers.

- Josh de Leeuw
- Jonny Saunders
- alexrockhill
