### 2024-03-25 #6512 @mstimberg
:wave: @magland Thank you for submitting to JOSS ! I will facilitate the review process as the editor, don't hesitate to ping me (@mstimberg) in case of any questions. The first step will be to find reviewers, if you have any suggestions, please mention them here (without tagging them with an `@`). Thanks!
