+++ python/neurosift/codecs/__init__.py
@@ -1,0 +1,1 @@
+from .MP4AVCCodec import MP4AVCCodec  # noqa: F401
