+++ python/setup.cfg
@@ -1,7 +1,7 @@
 # This file was automatically generated by jinjaroot. Do not edit directly.
 [metadata]
 name = neurosift
-version = 0.2.5
+version = 0.2.6
 author = Jeremy Magland
 author_email = jmagland@flatironinstitute.org
 description = Simple utility to view local NWB files using Neurosift
