### 2024-04-08 #6590 @editorialbot
[REVIEW]: Neurosift: DANDI exploration and NWB visualization in the browser
**Submitting author:** @magland (<a href="http://orcid.org/0000-0002-5286-4375">Jeremy Magland</a>)
**Repository:** https://github.com/flatironinstitute/neurosift
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.8
