### 2024-04-08 #6590 @mstimberg
👋🏼 @magland @alexrockhill, @jodeleeuw, @sanjayankur31 this is the review thread for the paper. All of our communications will happen here from now on.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/6590#issue-2230926089).

Please feel free to ping me (@mstimberg) if you have any questions/concerns. Looking forward to a constructive review!
