### 2024-04-08 #6590 @mstimberg
Please ignore the [statement about the licence](https://github.com/openjournals/joss-reviews/issues/6590#issuecomment-2042484912) and the [lines of code](https://github.com/openjournals/joss-reviews/issues/6590#issuecomment-2042484748) in the comments above. Those were run in the `paper` branch, which doesn't contain code and licence.
As correctly detected in the [pre-review](https://github.com/openjournals/joss-reviews/issues/6512#issuecomment-2012686570), the licence is an OSI-approved Apache 2.0 licence. If you are curious, you can find the LOC count in the pre-review issue as well: https://github.com/openjournals/joss-reviews/issues/6512#issuecomment-2012685994
