### 2024-04-16 #6590 @alexrockhill
I'm reviewing the functionality by selecting a few random datasets. Am I supposed to be able to interactively visualize the time series for this dataset https://neurosift.app/?p=/nwb&dandisetId=000122&dandisetVersion=draft&url=https://api.dandiarchive.org/api/assets/37f5e3ce-e1b7-48a7-a47d-6745875eacaa/download/&tab=neurodata-item:/stimulus/presentation/auditory|TimeSeries? I don't see anything in the manuscript about ElectricalTimeSeries and Auditory TimeSeries so this might fall under not yet implemented but it's somewhat similar to the Spike Raster Plot in Figure 3 so even if it is outside of the scope, it would be nice to have within the manuscript a table or some such with what kinds of data are implemented and which are on the roadmap or perhaps even not planned to be implemented.
