### 2024-04-16 #6590 @alexrockhill
Are the contents of .vscode cruft? (https://github.com/flatironinstitute/neurosift/blob/main/.vscode) Or do they serve some purpose for the software? I'm guessing these could maybe be removed.
