### 2024-04-16 #6590 @alexrockhill
I don't see any tests and it looks like they are not integrated into the CIs of pull requests based on the merged pull requests. This is a requirement for the review

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
