### 2024-04-16 #6590 @magland
> > Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> Just following the letter of the review guidelines here, could you please add instructions specific to where to 3) seek support? Whether that's through the github discussions or some other channel

There is a simple "Seeking Help" section [in the main README](https://github.com/flatironinstitute/neurosift/blob/main/README.md).
