### 2024-04-16 #6590 @magland
> Are the contents of .vscode cruft? (https://github.com/flatironinstitute/neurosift/blob/main/.vscode) Or do they serve some purpose for the software? I'm guessing these could maybe be removed.

One of the tasks was not being used, and so I removed that just now. The other two commands are used for incrementing the version of the Python package and deploying to pypi -- so those are still needed.
