### 2024-04-16 flatironinstitute/neurosift#149 @magland
TimeSeries View: Display helpful message when TimeSeries object is not numeric
See

https://neurosift.app/?p=/nwb&dandisetId=000122&dandisetVersion=draft&url=https://api.dandiarchive.org/api/assets/37f5e3ce-e1b7-48a7-a47d-6745875eacaa/download/&tab=neurodata-item:/stimulus/presentation/auditory|TimeSeries

and

https://github.com/openjournals/joss-reviews/issues/6590#issuecomment-2058015724
