### 2024-04-16 flatironinstitute/neurosift#149 @magland
This has been implemented.

![image](https://github.com/flatironinstitute/neurosift/assets/3679296/c105ed82-6bfb-4dad-9fd3-e510e1055256)
