### 2024-04-16 #6590 @alexrockhill
> There is a simple "Seeking Help" section [in the main README](https://github.com/flatironinstitute/neurosift/blob/main/README.md).

I saw that and looks great, except that if someone has a question like "does Neurosift do x?" or "how can I do x with Neurosift?" the community standard is not to have those type of questions as issues (issues being for bugs and enhancements). There's a newish GitHub discussions that is an easy option for this. I think ideally you would not want someone posting to issues seeking help so it would be great to clarify. Other projects like jsPsych use Github discussions https://www.jspsych.org/7.3/support/support/ or Discourse is another option e.g.
https://mne.discourse.group/ or a mailing list is the most classic.
