### 2024-04-16 #6590 @alexrockhill
> You are right, there are no automated tests at this point. But the way I read the instructions, this is not strictly necessary. It does have "manual steps described so that the functionality of the software can be verified". The main README contains several ways that you can verify that the software is functioning. If you want to test that the software works for visualizing a local .nwb file, you could download a relatively small example [from here](https://dandiarchive.org/dandiset/000946/draft/files?location=sub-BH494&page=1). If it would be helpful I can update the readme to point to that example.

I think following the letter here it would be nice to have specific standard instructions for how to run through a test. If I were to add a PR for a feature for instance, how would I know that I hadn't broken anything?
