### 2024-04-16 #6590 @magland
> I think following the letter here it would be nice to have specific standard instructions for how to run through a test. If I were to add a PR for a feature for instance, how would I know that I hadn't broken anything?

Okay, I'll see what I can do!
