### 2024-04-17 #6590 @magland
> > I think following the letter here it would be nice to have specific standard instructions for how to run through a test. If I were to add a PR for a feature for instance, how would I know that I hadn't broken anything?

@alexrockhill I have added a [table of tests](https://neurosift.app/?p=/tests) for many of the view plugins. In the [developer instructions](https://github.com/flatironinstitute/neurosift/blob/main/doc/development_environment.md) there is now information on how to bring this table up for a local development server, which will allow you to see if you have broken anything as you improve the existing views or add new views.
