### 2024-04-18 #6590 @alexrockhill
I don't see the table of tests when I follow the instructions, unfortunately
![image](https://github.com/openjournals/joss-reviews/assets/13473576/c04455bc-209f-4d87-b480-a910cec360c2)
