### 2024-04-18 #6590 @magland
@alexrockhill I wonder if you have pulled the latest code in the submodule?

git pull --recurse-submodules
