### 2024-04-18 #6590 @alexrockhill
> @alexrockhill I wonder if you have pulled the latest code in the submodule?
>
> git pull --recurse-submodules

alexrockhill@james:~/Downloads/neurosift$ git pull --recurse-submodules
Fetching submodule gui/fi-sci
Already up to date.
