### 2024-04-18 #6590 @magland
Thanks so much @alexrockhill ! Your comments and suggestions have been really helpful for polishing the software and docs.

I fixed the issue with the tests page on the dev server. I had neglected to update the submodule pointer in the main repo.

> I think you should spell out that contributors should compare the output with that on the main branch in order to make sure that their changes didn't break anything although I understand that this is fairly obvious.

This is done.

> Lastly, in checking a few other datasets, should you be able to view this dataset [https://neurosift.app/?p=/nwb&dandisetId=000576&dandisetVersion=0.231010.1811&url=https://api.dandiarchive.org/api/assets/57485b5a-e259-4759-a17a-4bef1e7f3886/download/&tab=neurodata-item:/processing/ecephys/LFP/ecephys.ieeg|ElectricalSeries](https://neurosift.app/?p=/nwb&dandisetId=000576&dandisetVersion=0.231010.1811&url=https://api.dandiarchive.org/api/assets/57485b5a-e259-4759-a17a-4bef1e7f3886/download/&tab=neurodata-item:/processing/ecephys/LFP/ecephys.ieeg%7CElectricalSeries) relative to the events like in Figure 2 of the paper or is that not part of the functionality claims?

If I am understanding this correctly... yes you can create a synchronized view like this
https://neurosift.app/?p=/nwb&dandisetId=000576&dandisetVersion=0.231010.1811&url=https://api.dandiarchive.org/api/assets/57485b5a-e259-4759-a17a-4bef1e7f3886/download/&tab=neurodata-items:neurodata-item:/processing/ecephys/LFP/ecephys.ieeg|ElectricalSeries@neurodata-item:/intervals/trials|TimeIntervals&tab-time=0,360.80789424126243,95.9129592790732
I need to update the docs to clarify how you create this view -- you need to click two checkboxes and then click the "view 2 items" button on the left.
