### 2024-04-19 #6590 @alexrockhill
![image](https://github.com/openjournals/joss-reviews/assets/13473576/cea52a2f-45d4-44d9-b55a-cbf35ef4bd1b)

Yes, that's exactly what I was talking about, but, unfortunately, I don't see the data, it doesn't appear for some reason.
