### 2024-04-19 #6590 @magland
@alexrockhill When you click on my link you don't see what is in my screenshot? If you are not able to then I'm not sure how to explain that. But if you do see it... notice the message "Zoom in" too see the data. You can also press Shift+O (documented behind the question mark in the left panel) to override.
