### 2024-04-30 #6590 @alexrockhill
> @alexrockhill When you click on my link you don't see what is in my screenshot? If you are not able to then I'm not sure how to explain that. But if you do see it... notice the message "Zoom in" too see the data. You can also press Shift+O (documented behind the question mark in the left panel) to override.

Ah yes, my bad, I was zooming in and not seeing the data but I just had to zoom in more. It's a bit disconcerting not to have any data shown when you load up, although the message is helpful. Is there lazy loading and would it be doable enough to get a highly decimated version to show when it's zoomed out? This isn't necessary for functionality but I think it would be nice for design.
