### 2024-04-30 #6590 @alexrockhill
Let me know when you're able to add example links to https://github.com/flatironinstitute/neurosift/blob/main/doc/neurodata_types.md and I will check off all the items in my review. I just think the project needs this documentation in order for users to fully understand the functionality and for contributors to have a somewhat comprehensive list of examples to check that they didn't break when they add a PR.
