### 2024-04-30 #6590 @mstimberg
Thanks for the nudge @magland, I have indeed not been following as closely as I should. @alexrockhill, thanks for all your comments so far, and @sanjayankur31, looking forward to your review.
@jodeleeuw Do you have an estimate as to when you will be able to take a look at the paper/software?
