### 2024-05-01 1ef902d
update examples table with links
