### 2024-05-05 #6590 @jodeleeuw
For documentation, I think it would be helpful to have an entry point in the main README that describes what documentation is available and includes a suggestion about where to get started for undocumented features. I wasn't sure where to look to find documentation initially, and ultimately got there through the link to the developer setup instructions in the README.
