### 2024-05-06 #6590 @magland
> > You are right, there are no automated tests at this point. But the way I read the instructions, this is not strictly necessary. It does have "manual steps described so that the functionality of the software can be verified". The main README contains several ways that you can verify that the software is functioning. If you want to test that the software works for visualizing a local .nwb file, you could download a relatively small example [from here](https://dandiarchive.org/dandiset/000946/draft/files?location=sub-BH494&page=1). If it would be helpful I can update the readme to point to that example.
>
> I think this would be helpful to add the link to the file. I'm not a user of NWB, so I'm not the target audience here, but this did trip me up when trying to get started. I could imagine it being useful for folks who just want to quickly try the local version out but don't have a small file handy.

@jodeleeuw I have added the following to the README

If you do not have an NWB file but want to try out this functionality, you can download [one of these relative small files](https://dandiarchive.org/dandiset/000946/draft/files?location=sub-BH494&page=1) from DANDI.
