### 2024-05-06 #6590 @magland
> For documentation, I think it would be helpful to have an entry point in the main README that describes what documentation is available and includes a suggestion about where to get started for undocumented features. I wasn't sure where to look to find documentation initially, and ultimately got there through the link to the developer setup instructions in the README.

Thanks for the suggestion @jodeleeuw . I have reorganized the README a bit and there is a new section called

[Where to find documentation](https://github.com/flatironinstitute/neurosift/blob/main/README.md#where-to-find-documentation)

Let me know if you notice anything missing from there.
