+++ python/neurosift/local-file-access-js/readme.md
@@ -27,7 +27,7 @@ npm run start $NWB_DIR
 **Step 3: Open the neurosift web app in your browser and point to an nwb file**
 
 ```bash
-https://flatironinstitute.github.io/neurosift/?p=/nwb&url=http://localhost:61762/files/testing.nwb
+https://neurosift.app/?p=/nwb&url=http://localhost:61762/files/testing.nwb
 ```
 
 This will load the nwb file from `$NWB_DIR/testing.nwb`
@@ -44,4 +44,4 @@ PORT=12345 npm run start $NWB_DIR
 
 This server will expose the contents of files in the directory you specify to anyone who can access the server. This does not include hidden files (those starting with "."). This is not a problem if you are running the server on your local machine and only accessing it from your local machine. However, if you are running the server on a machine that is accessible from the internet, you should take precautions to ensure that only authorized users can access the server, if you have sensitive files.
 
-There is also a possibility that websites you visit could get read access to your files. However, the configuration will prevent this from happening unless the website is being served from the allowed domains (flatironinsitute.github.io and localhost:3000).
+There is also a possibility that websites you visit could get read access to your files. However, the configuration will prevent this from happening unless the website is being served from the allowed domains (neurosift.app and localhost:3000).
