### 2024-05-07 #6590 @magland
> Hi @magland , looks very good. I just had some trouble running a local instance to view local files. The rest looks good to me 👍

Thanks @sanjayankur31

>
> I wasn't able to view local files, for example, [this one](https://github.com/OpenSourceBrain/NWBShowcase/raw/master/FergusonEtAl2015/FergusonEtAl2015.nwb) :
> ...

This was a CORS issue. It should be fixed now, with neurosift on pypi at version 0.2.8. I verified it worked with the following command using chrome and firefox

```bash
neurosift view-nwb FergusonEtAl2015.nwb
```

>
> I wasn't able to view the file using the live site either. This didn't quite seem to work:
>
> https://neurosift.app/?p=/nwb&url=https://github.com/OpenSourceBrain/NWBShowcase/raw/master/FergusonEtAl2015/FergusonEtAl2015.nwb
>
> The page seems to stop at `Loading https://github.com/OpenSourceBrain/NWBShowcase/raw/master/FergusonEtAl2015/FergusonEtAl2015.nwb`
>

This is also a CORS issue. While there is no way to get that particular URL to work due to the github server configuration being beyond our control, this other URL to the same file appears to work

https://neurosift.app/?p=/nwb&url=https://raw.githubusercontent.com/OpenSourceBrain/NWBShowcase/master/FergusonEtAl2015/FergusonEtAl2015.nwb

> In addition to NWB Widgets, it's also worth looking at [NWB Explorer](https://github.com/MetaCell/nwb-explorer) that aims to do something similar. It does not include streaming support at the moment, though.

Thank you. I have added this to a new "See also" section in the readme.

>
> General:
>
> * has there been an attempt to contribute the improvements to h5wasm back to upstream so that the benefits are available to the larger community too?

I agree that this could be beneficial to the larger community, and I have explored this possibility. Unfortunately my modifications were to the emscripten generated code, and so if this were to be done properly, the PR would need to go to [emscripten](https://emscripten.org/), which would require a lot more effort than I am prepared to expend at this point. Very low-level modifications were required to achieve efficient reading of chunks from the remote file.
