### 2024-05-14 #6590 @mstimberg
Many thanks @alexrockhill, @jodeleeuw, and @sanjayankur31 for your detailed and constructive reviews. It seems that you all recommend the publication of the software/paper in its current form (if not, please let me know :smile: ). @sanjayankur31, one item in your checklist is still unchecked, but I take your earlier comment as indicating that you are happy with everything.

@magland The next steps are some final checks by myself and preparing everything for the actual publication process. I will create a checklist in the next comment with some action items for you (and some more for me).
