### 2024-05-14 #6590 @sanjayankur31
> Many thanks @alexrockhill, @jodeleeuw, and @sanjayankur31 for your detailed and constructive reviews. It seems that you all recommend the publication of the software/paper in its current form (if not, please let me know 😄 ).

Yes :)

> @sanjayankur31, one item in your checklist is still unchecked, but I take your earlier comment as indicating that you are happy with everything.

Ah, sorry about that. Checked off now too.
