### 2024-05-14 #6590 @magland
> * Double check authors and affiliations (including ORCIDs)

I have double checked these.

> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

The latest version is 0.2.8

> * Archive the release on Zenodo/figshare/etc and post the DOI here.

Archived on Zenodo: https://zenodo.org/records/11192864

DOI: `10.5281/zenodo.11192864`

Note that I did not create an archive release in the traditional way because the source contains a git submodule. So I prepared a .tar.gz of the source code including the submodule and uploaded that to Zenodo.

> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

I double checked this.

> * Make sure that the license listed for the archive is the same as the software license.

It does match (Apache 2.0)
