### 2024-05-17 #6590 @mstimberg
> Note that I did not create an archive release in the traditional way because the source contains a git submodule. So I prepared a .tar.gz of the source code including the submodule and uploaded that to Zenodo.

Thanks, the archive looks good to me. While it is not strictly necessary, it would be preferable to also have a corresponding tag/release in the github repo. Note that you can also manually attach your `.tar.gz` file to the release.
