### 2024-05-17 ca9e83e
Add DOIs to bibliography and fix capitalization
