+++ paper/joss/paper.bib
@@ -1,21 +1,23 @@
 @article{rubel2022neurodata,
-  title={The neurodata without borders ecosystem for neurophysiological data science},
+  title={The {Neurodata Without Borders} ecosystem for neurophysiological data science},
   author={R{\"u}bel, Oliver and Tritt, Andrew and Ly, Ryan and Dichter, Benjamin K and Ghosh, Satrajit and Niu, Lawrence and Baker, Pamela and Soltesz, Ivan and Ng, Lydia and Svoboda, Karel and others},
-  journal={Elife},
+  journal={eLife},
   volume={11},
   pages={e78362},
   year={2022},
+  doi={10.7554/eLife.78362},
   publisher={eLife Sciences Publications Limited}
 }
 
 @article{teeters2015neurodata,
-  title={Neurodata without borders: creating a common data format for neurophysiology},
+  title={{Neurodata Without Borders}: creating a common data format for neurophysiology},
   author={Teeters, Jeffery L and Godfrey, Keith and Young, Rob and Dang, Chinh and Friedsam, Claudia and Wark, Barry and Asari, Hiroki and Peron, Simon and Li, Nuo and Peyrache, Adrien and others},
   journal={Neuron},
   volume={88},
   number={4},
   pages={629--634},
   year={2015},
+  doi={10.1016/j.neuron.2015.10.025},
   publisher={Elsevier}
 }
 
@@ -25,7 +27,7 @@
   year = {2022},
   publisher = {GitHub},
   journal = {GitHub repository},
-  howpublished = {\url{https://github.com/NeurodataWithoutBorders/nwbwidgets}},
+  url = {https://github.com/NeurodataWithoutBorders/nwbwidgets},
   note = {Accessed: 2024-02-26},
 }
 
@@ -35,7 +37,7 @@
   year = {2022},
   publisher = {GitHub},
   journal = {GitHub repository},
-  howpublished = {\url{https://github.com/usnistgov/h5wasm}},
+  url = {https://github.com/usnistgov/h5wasm},
   note = {Accessed: 2024-02-26}
 }
 
@@ -45,6 +47,6 @@
   year = {2024},
   publisher = {Zenodo},
   journal = {Zenodo},
-  howpublished = {\url{https://doi.org/10.5281/zenodo.10790679}},
+  doi = {10.5281/zenodo.10790679},
   note = {Accessed: 2024-03-15}
 }
