### 2024-05-17 57e07e8
Put class names in italics everywhere
