+++ paper/joss/paper.md
@@ -46,13 +46,13 @@ Because files found on DANDI can often be large and unwieldy, various tools have
 Neurodata Without Borders files are structured hierarchically and encapsulate various "neurodata" types that reflect different aspects of neurophysiological experiments. These types range from *BehavioralEvents*, which record discrete actions or occurrences within experiments, to data structures like *Fluorescence*, *ImageSegmentation*, and *RoiResponseSeries*, which are key data types in optical neurophysiology. Other neurodata types include *ElectricalSeries* for electrophysiological signals and *Units* for spike times of neurons. Neurosift allows interactive navigation of this hierarchical structure (Figure 1) and provides plugin visualizations for many of these types (Figure 2). It also facilitates the creation of composite views by allowing users to select and synchronize multiple data types within the same interface (Figure 3). This synchronization extends to navigation actions such as zooming and panning, where different sub-windows, each displaying a different aspect of the data, maintain a shared time axis. These views can then be shared with others as a URL.
 
 <!-- https://neurosift.app/?p=/nwb&url=https://api.dandiarchive.org/api/assets/db8b5f00-05a1-4485-89ca-1767acf0ae36/download/&dandisetId=000409&dandisetVersion=draft -->
-![Screenshot of Neurosift displaying a NWB file from the 'IBL - Brain Wide Map' Dandiset (ID 000409), showcasing the tool's capability to navigate and visualize the hierarchical structure of neurophysiological data. Visible are expanded panels for ElectricalSeries and ImageSeries objects, alongside a TimeIntervals object and a Units table.](./images/neurosift-screenshot-1.png)
+![Screenshot of Neurosift displaying a NWB file from the 'IBL - Brain Wide Map' Dandiset (ID 000409), showcasing the tool's capability to navigate and visualize the hierarchical structure of neurophysiological data. Visible are expanded panels for *ElectricalSeries* and *ImageSeries* objects, alongside a *TimeIntervals* object and a *Units* table.](./images/neurosift-screenshot-1.png)
 
 <!-- https://neurosift.app/?p=/nwb&url=https://dandiarchive.s3.amazonaws.com/blobs/f76/b9f/f76b9f7e-38ed-43fb-8fd0-4cc272fcd8bd&dandisetId=000409&dandisetVersion=draft&tab=view:PSTH|/intervals/trials -->
 ![Interactive Peri-stimulus Time Histogram (PSTH) visualization in Neurosift, enabling users to select neural units, time alignment variables, and customize options like window range, histogram bin count, and grouping variables.](./images/neurosift-screenshot-3.png)
 
 <!-- https://neurosift.app/?p=/nwb&url=https://dandiarchive.s3.amazonaws.com/blobs/c86/cdf/c86cdfba-e1af-45a7-8dfd-d243adc20ced&tab=neurodata-items:neurodata-item:/acquisition/ElectricalSeries|ElectricalSeries@view:DirectRasterPlot|/units -->
-![Synchronized view within Neurosift demonstrating the interactive alignment between an ElectricalSeries and a Spike Raster Plot. This feature allows users to seamlessly zoom and pan across both visualizations, maintaining a coherent temporal perspective across different data types.](./images/neurosift-screenshot-2.png)
+![Synchronized view within Neurosift demonstrating the interactive alignment between an *ElectricalSeries* and a *Spike Raster Plot*. This feature allows users to seamlessly zoom and pan across both visualizations, maintaining a coherent temporal perspective across different data types.](./images/neurosift-screenshot-2.png)
 
 
 
