### 2024-05-17 #6590 @bendichter
Great work all-around. The paper is quite well-written, and the review process was very impressive. The proof looks good to me!

NWB Widgets does not currently have a DOI. If that is critical I can mint one on Zenodo, though I am fine with leaving the citation as-is.
