### 2024-05-25 #6590 @Kevin-Mattheus-Moerman
- [ ] In your affiliations, please spell out `USA` as `United States of America`
- [ ] Please set `0.2.8` as the version tag on ZENODO
- [ ] Please create a tagged release on your repository with the version tag `0.2.8`.
