### 2024-05-27 #6590 @magland
> @magland as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, however the below are some points that require your attention:
>
> * [ ]  In your affiliations, please spell out `USA` as `United States of America`

Done
https://github.com/flatironinstitute/neurosift/blob/paper/paper/joss/paper.md

> * [ ]  Please set `0.2.8` as the version tag on ZENODO

Done
https://zenodo.org/records/11192864

> * [ ]  Please create a tagged release on your repository with the version tag `0.2.8`.

Done
https://github.com/flatironinstitute/neurosift/releases/tag/v0.2.8

@Kevin-Mattheus-Moerman thanks so much!

Let me know if there's anything else that needs to be done.
