### 2024-05-27 #6590 @magland
> @magland please edit the Zenodo listed version to include the v in v0.2.8

@Kevin-Mattheus-Moerman done
