### 2024-05-27 18493ce
Create CITATION.cff
