### 2024-05-28 #6590 @Kevin-Mattheus-Moerman
Congratulations on this JOSS publication @magland ! :partying_face:

Thanks for editing @mstimberg !

And a special thank you to the reviewers: @alexrockhill, @jodeleeuw, @sanjayankur31 !!!!
